[Mission] improve dates error displaying
